+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Bugfix
 - Documentation : Fixes in the rename of PHPPowerPoint - @Progi1984 GH-127
 - ODPresentation : Exclude SVM files for reader - @Progi1984 GH-141
+- PowerPoint2007 Writer : Bugfix for opening PPTX on Mac - @thsteinmetz GH-89
 
 ### Changes
 - PhpOffice\PhpPowerpoint\Style\Alignment::setLevel can now be defined great than 8 - @Progi1984 GH-141